+++ tests/callbacks/test_aggregate_predictions.py
@@ -1,5 +1,7 @@
 from unittest import TestCase
 import torch
+import warnings
+
 from torchbearer.callbacks import AggregatePredictions
 import torchbearer
 
@@ -44,4 +46,16 @@ class TestAggregatePredictions(TestCase):
         aggregator.on_end_epoch(state_2)
         self.assertTrue(list(aggregator.predictions_list) == [])
 
+    def test_none_predictions(self):
+        aggregator = AggregatePredictions()
 
+        with warnings.catch_warnings(record=True) as w:
+            state_1 = {torchbearer.Y_PRED: [None]}
+
+            aggregator.on_step_validation(state_1)
+            aggregator.on_step_validation(state_1)
+
+            self.assertTrue(list(aggregator.predictions_list) == [[None], [None]])
+
+            aggregator.on_end_validation(state_1)
+            self.assertTrue(state_1[torchbearer.FINAL_PREDICTIONS] == [[None], [None]])