+++ torchbearer/callbacks/aggregate_predictions.py
@@ -1,6 +1,7 @@
 import torchbearer
 from torchbearer.callbacks import Callback
 import torch
+import warnings
 
 
 class AggregatePredictions(Callback):
@@ -13,9 +14,14 @@ class AggregatePredictions(Callback):
         self.predictions_list.append(state[torchbearer.Y_PRED])
 
     def on_end_validation(self, state):
-
-        state[torchbearer.FINAL_PREDICTIONS] = torch.cat(self.predictions_list, 0) if len(self.predictions_list) != 1 \
-            else self.predictions_list[0]
+        try:
+            if len(self.predictions_list) == 1 and type(self.predictions_list[0]) is torch.Tensor:
+                state[torchbearer.FINAL_PREDICTIONS] = self.predictions_list[0]
+            else:
+                state[torchbearer.FINAL_PREDICTIONS] = torch.cat(self.predictions_list, 0)
+        except:
+            warnings.warn('Failed to format predictions as tensor, returning as list.')
+            state[torchbearer.FINAL_PREDICTIONS] = self.predictions_list
 
     def on_end_epoch(self, state):
         super(AggregatePredictions, self).on_end_epoch(state)