+++ CHANGELOG.md
@@ -3,6 +3,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased]
+### Added
+### Changed
+### Deprecated
+### Removed
+### Fixed
+- Fixed bug where aggregate predictions couldn't handle empty list
+
 ## [0.4.0] - 2019-07-05
 ### Added
 - Added ``with_loader`` trial method that allows running of custom batch loaders